+++ test-utils/CHANGELOG.md
@@ -3,3 +3,5 @@
 All notable changes to this package will be documented in this file. Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+- Added `golden` package: minimal golden-file test harness with -update flag and GOLDEN_UPDATE env support (#104).